# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent AI System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/discovery/agents": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Discover Agents",
        "description": "Search for agents based on a text query using semantic search.\n\nReturns a list of A2A-compliant AgentCards for agents that match the query\nwith a similarity score above the confidence threshold.\n\n**Authentication:** Requires `X-API-Key` header with a valid API key.\n\n**Example Request:**\n```json\n{\n    \"query\": \"I need help with data analysis\",\n    \"limit\": 5\n}\n```\n\n**Example Response:**\n```json\n{\n    \"query\": \"I need help with data analysis\",\n    \"agents\": [\n        {\n            \"agent_card\": { ... },\n            \"match_score\": 0.92\n        }\n    ],\n    \"count\": 1\n}\n```",
        "operationId": "discover_agents_api_v1_discovery_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys": {
      "get": {
        "tags": [
          "api_keys"
        ],
        "summary": "List API Keys",
        "description": "List all API keys belonging to the authenticated user.",
        "operationId": "list_api_keys_api_v1_api_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "api_keys"
        ],
        "summary": "Create API Key",
        "description": "Create a new API key and return its plaintext value once.",
        "operationId": "create_api_key_api_v1_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "api_keys"
        ],
        "summary": "Deactivate API Key",
        "description": "Soft-delete an API key owned by the authenticated user.",
        "operationId": "deactivate_api_key_api_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyOperationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Friendly name for the API key"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "APIKeyCreateRequest"
      },
      "APIKeyCreateResponse": {
        "properties": {
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key"
          }
        },
        "type": "object",
        "required": [
          "key_id",
          "name",
          "created_at",
          "api_key"
        ],
        "title": "APIKeyCreateResponse"
      },
      "APIKeyErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "APIKeyErrorResponse"
      },
      "APIKeyItemResponse": {
        "properties": {
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count"
          }
        },
        "type": "object",
        "required": [
          "key_id",
          "name",
          "created_at",
          "is_active",
          "usage_count"
        ],
        "title": "APIKeyItemResponse"
      },
      "APIKeyListResponse": {
        "properties": {
          "keys": {
            "items": {
              "$ref": "#/components/schemas/APIKeyItemResponse"
            },
            "type": "array",
            "title": "Keys"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "keys",
          "count"
        ],
        "title": "APIKeyListResponse"
      },
      "APIKeyOperationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "APIKeyOperationResponse"
      },
...
      "DiscoveryAgentResult": {
        "properties": {
          "agent_card": {
            "additionalProperties": true,
            "type": "object",
            "title": "Agent Card"
          },
          "match_score": {
            "type": "number",
            "title": "Match Score"
          }
        },
        "type": "object",
        "required": [
          "agent_card",
          "match_score"
        ],
        "title": "DiscoveryAgentResult",
        "description": "A single agent result from the Discovery API."
      },
      "DiscoveryErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "DiscoveryErrorResponse",
        "description": "Standardized error response for Discovery API."
      },
      "DiscoveryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Text query to search for matching agents"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100,
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit",
            "description": "Maximum number of agents to return (default: 10, max: 100)"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DiscoveryRequest",
        "description": "Request body for agent discovery endpoint."
      },
      "DiscoveryResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/DiscoveryAgentResult"
            },
            "type": "array",
            "title": "Agents"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "query",
          "agents",
          "count"
        ],
        "title": "DiscoveryResponse",
        "description": "Successful response from the Discovery API."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}